+++ backend/dal/model/items.gen.go
@@ -12,7 +12,7 @@ const TableNameItem = "items"
 
 // Item mapped from table <items>
 type Item struct {
-	ID        int32     `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID        int64     `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	Name      string    `gorm:"column:name" json:"name"`
 	CreatedAt time.Time `gorm:"column:created_at;default:CURRENT_TIMESTAMP" json:"created_at"`
 	UpdatedAt time.Time `gorm:"column:updated_at;default:CURRENT_TIMESTAMP" json:"updated_at"`
